Renamed token.py into saml_token.py
